Text: separate glyph ID and offset/advance retrieval in AbstractShaper.
In basically all cases it's two independent operations so forcing them
to be done together doesn't really bring any potential efficiency
advantages. On the other hand, splitting them allows allows the caller
to better make use of available memory, as the new
renderGlyphQuadsInto() allows the input and output arrays to be aliased.

Bumping AbstractFont plugin interface versions as this is a breaking
change.