GL: add a repro case for base/range buffer unbind state mismatch.
Expands the test added in 789c52fd8af309b98801b4ced74260cef3d272de, for
which a fix was done in 8f6f4053fcd7893511a748df35e9cda881d390bf but
which forgot to handle the case where a buffer is unbound. The test now
fails with an invalid error.