Platform: make Sdl2Application::setMinimalLoopPeriod() unit-safe.
I wanted to add this for GlfwApplication, only to realize the timer there
is a double, in seconds, so accepting *integral* milliseconds there felt
very weird. Let's use the fancy new time types instead.

Also updated the docs to (hopefully) clarify that setSwapInterval() has
to be called in order for the interaction between the two to work
properly.

As usual, the old variant taking untyped milliseconds is a deprecated
alias to this one.