Trade: redo MeshData class-level documentation to make more sense.
Back in 2020 when I wrote this I didn't really expect the MeshData to be
directly used for much more than putting them on a GPU, mostly because
that used to be the primary use case with the old MeshData2D /
MeshData3D. So the documentation was focusing mainly on populating a GPU
mesh, and any docs for CPU-side access were added rather hastily.

As now the asset processing use case is much larger, the original docs no
longer made sense. Let's hope this is better.