Assert non-implementation-specific PixelFormat in Image* constructors.
And document that. Because the pixel size cannot be determined for it,
and one has to either pass it explicitly or use the templated overload
that figures it out implicitly via ADL. This asserted before, but only
deep inside in pixelFormatSize(), which may be confusing.

I need to do a similar treatment for compressed images with block size
properties so let's first make it behave properly for uncompressed.